requirements.txt
opencv-python-headless
torch
ultralytics
easyocr
numpy
Pillow
pytesseract